Changed packet_id type from 64 to 32 bit int
32 bits is already overkill, 64 bits is far beyond that

+++ src/PC_Network.hpp
@@ -50,7 +50,7 @@
  */
 struct AudioPacket
 {
-	uint64_t packet_id;
+	uint32_t packet_id;
 	uint16_t packet_len;
 	std::unique_ptr<uint8_t> packet;
